Rename folders and remove unnecessary packages from Auth
