+++ Auth/package.json
@@ -12,16 +12,11 @@
     "license": "ISC",
     "dependencies": {
         "-": "^0.0.1",
-        "bcrypt": "^5.1.1",
         "body-parser": "^1.20.2",
         "cors": "^2.8.5",
         "dotenv": "^16.3.1",
         "express": "^4.18.2",
-        "jsonwebtoken": "^9.0.2",
-        "mongodb": "^6.0.0",
-        "mongoose": "^7.5.0",
-        "mysql2": "^3.6.1",
-        "save": "^2.9.0"
+        "jsonwebtoken": "^9.0.2"
     },
     "devDependencies": {
         "nodemon": "^3.0.1"
